+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Fix healing a list with a default scope `:order` and/or `:select`. Thanks @LukasSkywalker!
+
 ## [0.4.4] - 2024-11-20
 
 - Add `funding_uri` to gemspec.